added martysama files modifications

+++ source/server/game/src/messenger_manager.cpp
@@ -12,6 +12,11 @@
 #include "char_manager.h"
 #include "questmanager.h"
 
+// @fixme142 BEGIN
+static char	__account[CHARACTER_NAME_MAX_LEN*2+1];
+static char	__companion[CHARACTER_NAME_MAX_LEN*2+1];
+// @fixme142 END
+
 MessengerManager::MessengerManager()
 {
 }
@@ -43,8 +48,14 @@ void MessengerManager::Login(MessengerManager::keyA account)
 	if (m_set_loginAccount.find(account) != m_set_loginAccount.end())
 		return;
 
+	// @fixme142 BEGIN
+	DBManager::instance().EscapeString(__account, sizeof(__account), account.c_str(), account.size());
+	if (account.compare(__account))
+		return;
+	// @fixme142 END
+
 	DBManager::instance().FuncQuery(std::bind(&MessengerManager::LoadList, this, std::placeholders::_1),
-			"SELECT account, companion FROM messenger_list%s WHERE account='%s'", get_table_postfix(), account.c_str());
+			"SELECT account, companion FROM messenger_list%s WHERE account='%s'", get_table_postfix(), __account);
 
 	m_set_loginAccount.insert(account);
 }
@@ -138,7 +149,8 @@ void MessengerManager::RequestToAdd(LPCHARACTER ch, LPCHARACTER target)
 	target->ChatPacket(CHAT_TYPE_COMMAND, "messenger_auth %s", ch->GetName());
 }
 
-void MessengerManager::AuthToAdd(MessengerManager::keyA account, MessengerManager::keyA companion, bool bDeny)
+// @fixme130 void -> bool
+bool MessengerManager::AuthToAdd(MessengerManager::keyA account, MessengerManager::keyA companion, bool bDeny)
 {
 	DWORD dw1 = GetCRC32(companion.c_str(), companion.length());
 	DWORD dw2 = GetCRC32(account.c_str(), account.length());
@@ -150,7 +162,7 @@ void MessengerManager::AuthToAdd(MessengerManager::keyA account, MessengerManage
 	if (m_set_requestToAdd.find(dwComplex) == m_set_requestToAdd.end())
 	{
 		sys_log(0, "MessengerManager::AuthToAdd : request not exist %s -> %s", companion.c_str(), account.c_str());
-		return;
+		return false;
 	}
 
 	m_set_requestToAdd.erase(dwComplex);
@@ -160,6 +172,7 @@ void MessengerManager::AuthToAdd(MessengerManager::keyA account, MessengerManage
 		AddToList(companion, account);
 		AddToList(account, companion);
 	}
+	return true;
 }
 
 void MessengerManager::__AddToList(MessengerManager::keyA account, MessengerManager::keyA companion)
@@ -191,9 +204,16 @@ void MessengerManager::AddToList(MessengerManager::keyA account, MessengerManage
 	if (m_Relation[account].find(companion) != m_Relation[account].end())
 		return;
 
+	// @fixme142 BEGIN
+	DBManager::instance().EscapeString(__account, sizeof(__account), account.c_str(), account.size());
+	DBManager::instance().EscapeString(__companion, sizeof(__companion), companion.c_str(), companion.size());
+	if (account.compare(__account) || companion.compare(__companion))
+		return;
+	// @fixme142 END
+
 	sys_log(0, "Messenger Add %s %s", account.c_str(), companion.c_str());
-	DBManager::instance().Query("INSERT INTO messenger_list%s VALUES ('%s', '%s')", 
-			get_table_postfix(), account.c_str(), companion.c_str());
+	DBManager::instance().Query("INSERT INTO messenger_list%s VALUES ('%s', '%s')",
+			get_table_postfix(), __account, __companion);
 
 	__AddToList(account, companion);
 
@@ -222,9 +242,16 @@ void MessengerManager::RemoveFromList(MessengerManager::keyA account, MessengerM
 	if (companion.size() == 0)
 		return;
 
+	// @fixme142 BEGIN
+	DBManager::instance().EscapeString(__account, sizeof(__account), account.c_str(), account.size());
+	DBManager::instance().EscapeString(__companion, sizeof(__companion), companion.c_str(), companion.size());
+	if (account.compare(__account) || companion.compare(__companion))
+		return;
+	// @fixme142 END
+
 	sys_log(1, "Messenger Remove %s %s", account.c_str(), companion.c_str());
 	DBManager::instance().Query("DELETE FROM messenger_list%s WHERE account='%s' AND companion = '%s'",
-			get_table_postfix(), account.c_str(), companion.c_str());
+			get_table_postfix(), __account, __companion);
 
 	__RemoveFromList(account, companion);
 
@@ -240,9 +267,15 @@ void MessengerManager::RemoveAllList(keyA account)
 {
 	std::set<keyT>	company(m_Relation[account]);
 
+	// @fixme142 BEGIN
+	DBManager::instance().EscapeString(__account, sizeof(__account), account.c_str(), account.size());
+	if (account.compare(__account))
+		return;
+	// @fixme142 END
+
 	/* SQL Data 삭제 */
 	DBManager::instance().Query("DELETE FROM messenger_list%s WHERE account='%s' OR companion='%s'",
-			get_table_postfix(), account.c_str(), account.c_str());
+			get_table_postfix(), __account, __account);
 
 	/* 내가 가지고있는 리스트 삭제 */
 	for (std::set<keyT>::iterator iter = company.begin();
